Make file naming consistent and move default welo util to tests.
