+++ test/test-database.ts
@@ -13,7 +13,7 @@ import { Manifest } from '@/manifest/index.js'
 import { Blocks } from '@/blocks/index.js'
 
 import getDatastore from './utils/level-datastore.js'
-import defaultManifest from './utils/defaultManifest.js'
+import defaultManifest from './utils/default-manifest.js'
 import { getTestPaths, tempPath } from './utils/constants.js'
 import { getTestIpfs, offlineIpfsOptions } from './utils/ipfs.js'
 import { getTestIdentities, getTestIdentity } from './utils/identities.js'
