+++ test/utils/default-manifest.ts
